Muevo decorators fuera de la clase de test
refs:#87
